Bug 1776658: Update some setNumber callers. r=jandem
Differential Revision: https://phabricator.services.mozilla.com/D152822

UltraBlame original commit: 88c94f32cf0271a185cfc7a47665d3be5452e858

+++ dom/indexedDB/KeyPath.cpp
@@ -87,7 +87,7 @@ nsresult GetJSValFromKeyPathString(
       
       if (currentVal.isString() && !tokenizer.hasMoreTokens() &&
           token.EqualsLiteral("length")) {
-        aKeyJSVal->setNumber(double(JS_GetStringLength(currentVal.toString())));
+        aKeyJSVal->setNumber(JS_GetStringLength(currentVal.toString()));
         break;
       }
